create ProposalDescription component
