+++ src/components/Proposal/ProposalDetailsPage.tsx
@@ -29,8 +29,7 @@ import VoteButtons from "./Voting/VoteButtons";
 import VoteGraph from "./Voting/VoteGraph";
 import VotersModal from "./Voting/VotersModal";
 import * as css from "./ProposalDetails.scss";
-
-const ReactMarkdown = require("react-markdown");
+import ProposalDescription from "components/Shared/ProposalDescription";
 
 interface IExternalProps extends RouteComponentProps<any> {
   currentAccountAddress: Address;
@@ -113,54 +112,6 @@ class ProposalDetailsPage extends React.Component<IProps, IState> {
     this.setState({ showVotersModal: false });
   }
 
-  private parseYouTubeVideoIdFromUri = (url: string): string => {
-    const match = url.match(/(\/|%3D|v=)([0-9A-z-_]{11})([%#?&]|$)/);
-    if (match) {
-      if (match.length >= 3) {
-        return match[2];
-      } else {
-        // eslint-disable-next-line no-console
-        console.error("The outube url is not valid.");
-      }
-    }
-    return null;
-  }
-
-  private getVimeoIdFromUrl = (url: string): string => {
-    const match = url.match(/^.*(?:vimeo.com)\/(?:channels\/|channels\/\w+\/|groups\/[^/]*\/videos\/|album\/\d+\/video\/|video\/|)(\d+)(?:$|\/|\?)/);
-    if (match) {
-      if (match.length >= 2) {
-        return match[1];
-      } else {
-      // eslint-disable-next-line no-console
-        console.error("The vimeo url is not valid.");
-      }
-    }
-    return null;
-  }
-
-  private renderDescription = (props: { href: string; children: React.ReactNode }) => {
-    if (props.href) {
-      const url = new URL(props.href);
-      const videoId = this.parseYouTubeVideoIdFromUri(props.href);
-      if (videoId) {
-        const start = url.searchParams.get("t") || "0";
-
-        return <iframe className={css.embeddedVideo} frameBorder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowFullScreen
-          src={`${url.protocol}//www.youtube-nocookie.com/embed/${videoId}?start=${start}`}>
-        </iframe>;
-      } else {
-        const videoId = this.getVimeoIdFromUrl(props.href);
-        if (videoId) {
-          return <iframe className={css.embeddedVideo} frameBorder="0" allow="autoplay; fullscreen" allowFullScreen
-            src={`${url.protocol}//player.vimeo.com/video/${videoId}`}>
-          </iframe>;
-        }
-      }
-    }
-    return <a href={props.href} target="_blank" rel="noopener noreferrer">{props.children}</a>;
-  }
-
   public render(): RenderOutput {
     const {
       beneficiaryProfile,
@@ -238,7 +189,7 @@ class ProposalDetailsPage extends React.Component<IProps, IState> {
             </div>
 
             <div className={css.description}>
-              <ReactMarkdown source={proposal.description} renderers={{ link: this.renderDescription}} />
+              <ProposalDescription description={proposal.description} />
             </div>
 
             {url ?
