+++ src/components/Shared/ProposalDescription.scss
@@ -1,0 +1,3 @@
+iframe.embeddedVideo {
+  margin-top: 10px;
+}
